backend/requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
Pillow==10.2.0
psycopg2-binary==2.9.9
pyjwt==2.8.0
django-storages==1.14.2
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0